Committing v3.4
- Improved install
- Support for go.mod
- Added better install testing
- Added some logging
- Added support for dependencies in sharpcd.yml
