+++ .gitignore
@@ -2,3 +2,4 @@
 .vscode
 __debug_bin
 docker
+installs/testing/
