+++ .version
@@ -1,1 +1,1 @@
-VERSION=3.3
+VERSION=3.4
