+++ go.mod
@@ -1,0 +1,22 @@
+module sharpCD
+
+go 1.18
+
+require (
+	github.com/common-nighthawk/go-figure v0.0.0-20210622060536-734e95fb86be
+	github.com/gizak/termui/v3 v3.1.0
+	github.com/gorilla/websocket v1.5.0
+	github.com/hashicorp/go-version v1.5.0
+	github.com/joho/godotenv v1.4.0
+	golang.org/x/crypto v0.0.0-20220525230936-793ad666bf5e
+	gopkg.in/yaml.v2 v2.4.0
+)
+
+require (
+	github.com/mattn/go-runewidth v0.0.13 // indirect
+	github.com/mitchellh/go-wordwrap v1.0.1 // indirect
+	github.com/nsf/termbox-go v1.1.1 // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
+	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
+	golang.org/x/term v0.0.0-20220526004731-065cf7ba2467 // indirect
+)
