--- gopack.yml
@@ -1,14 +1,0 @@
-goversion: "1.12"
-packages:
-- gopkg.in/yaml.v2
-- golang.org/x/crypto/bcrypt
-- github.com/joho/godotenv
-- golang.org/x/term
-- github.com/hashicorp/go-version
-- github.com/gosuri/uilive
-- github.com/gizak/termui
-- github.com/common-nighthawk/go-figure
-- github.com/gorilla/websocket
-
-devpackages:
-- golang.org/x/sys/unix
